rename: move files and fix typo
