+++ src/app/(home)/components/OnReservation.tsx
@@ -1,8 +1,8 @@
 import Article from '@/components/article/Article';
 
-const OnReservationSection = () => {
+const OnReservation = () => {
   return (
-    <Article richTitle="지금 예약 중인 셔틀" showMore="/TODO">
+    <Article richTitle="지금 예약 모집 중인 셔틀" showMore="/TODO">
       {/* <div className="w-full p-16">
         <AskHandyButton sub="찾고 있는 셔틀이 없나요?">
           원하는 셔틀 요청하기
@@ -12,4 +12,4 @@ const OnReservationSection = () => {
   );
 };
 
-export default OnReservationSection;
+export default OnReservation;
