+++ src/app/(home)/help/about/components/AboutService.tsx
@@ -2,9 +2,9 @@ import { type PropsWithChildren } from 'react';
 import { StaticImport } from 'next/dist/shared/lib/get-img-props';
 import Image from 'next/image';
 import Article from '@/components/article/Article';
-import Person1 from './images/person1.png';
-import Person2 from './images/person2.png';
-import Person3 from './images/person3.png';
+import Person1 from '../images/person1.png';
+import Person2 from '../images/person2.png';
+import Person3 from '../images/person3.png';
 
 const AboutService = () => (
   <Article
