+++ src/app/(home)/images/1-intro.png
